+++ bespin_api_v2/apps.py
@@ -1,0 +1,7 @@
+from __future__ import unicode_literals
+
+from django.apps import AppConfig
+
+
+class BespinApiV2Config(AppConfig):
+    name = 'bespin_api_v2'
