+++ bespin_api_v2/urls.py
@@ -1,0 +1,20 @@
+from data import api as data_api
+from bespin_api_v2 import api
+from django.conf.urls import url, include
+from rest_framework import routers
+
+router = routers.DefaultRouter()
+router.register(r'workflows', data_api.WorkflowsViewSet, 'workflow')
+router.register(r'workflow-versions', data_api.WorkflowVersionsViewSet, 'workflowversion')
+router.register(r'workflow-configurations', api.WorkflowConfigurationViewSet, 'workflowconfigurations')
+router.register(r'vm-strategies', api.VMStrategyViewSet, 'vmstrategies')
+router.register(r'jobs', data_api.JobsViewSet, 'job')
+router.register(r'job-file-stage-groups', data_api.JobFileStageGroupViewSet, 'jobfilestagegroup')
+router.register(r'dds-job-input-files', data_api.DDSJobInputFileViewSet, 'ddsjobinputfile')
+router.register(r'url-job-input-files', data_api.URLJobInputFileViewSet, 'urljobinputfile')
+router.register(r'dds-endpoints', data_api.DDSEndpointViewSet, 'ddsendpoint')
+router.register(r'dds-user-credentials', data_api.DDSUserCredViewSet, 'ddsusercredential')
+
+urlpatterns = [
+    url(r'^', include(router.urls)),
+]
