+++ data/serializers.py
@@ -383,33 +383,3 @@ class AdminImportWorkflowQuestionnaireSerializer(serializers.Serializer):
     share_group_name = serializers.CharField(min_length=1) # must relate to an existing Share Group
     volume_size_base = serializers.IntegerField()
     volume_size_factor = serializers.IntegerField()
-
-
-class VMStrategySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = VMStrategy
-        resource_name = 'vm-strategies'
-        fields = '__all__'
-
-
-class WorkflowConfigurationSerializer(serializers.ModelSerializer):
-    tag = serializers.SerializerMethodField()
-    user_fields_json = serializers.SerializerMethodField()
-
-    def get_tag(self, obj):
-        return obj.make_tag()
-
-    def get_user_fields_json(self, obj):
-        fields = json.loads(obj.workflow_version.fields_json)
-        system_order_json = json.loads(obj.system_job_order_json)
-        system_keys = system_order_json.keys()
-        user_fields_json = []
-        for field in fields:
-            if field['name'] not in system_keys:
-                user_fields_json.append(field)
-        return json.dumps(user_fields_json)
-
-    class Meta:
-        model = WorkflowConfiguration
-        resource_name = 'workflow-configuration'
-        fields = '__all__'
